% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{descriptive_groupby}
\alias{descriptive_groupby}
\title{Group wise descriptive summary of a numeric vector.}
\usage{
descriptive_groupby(x, group, xname = deparse(substitute(x)))
}
\arguments{
\item{x}{must be numeric.}

\item{group}{must be factor.}

\item{xname}{Name of numeric variable. Default is the name of x.}
}
\value{
N = Number of observations in \code{x}.

Total= Sum of \code{x}.

MEAN= Mean of \code{x}.

SD= Standard deviation of \code{x}.

MEDIAN= Median of \code{x}.

MIN= Minimum of \code{x}.

MAX= Maximum of \code{x}.

IQR= Interquartile range (Q1,Q3) of \code{x}.i.e., 1st quantile and the 3rd quantile.

Note: All the summary statistics will be calculated after removing the \code{NA} values.
}
\description{
To generate summary statistics for a numerical variable '\code{x}' at different levels of a particular factor variable '\code{group}' in a \code{data.frame}.
}
\examples{

## IRIS data:

# descriptive summary of  Sepal Length, Sepal Width, Petal Length and, Petal Width

rbind(with(iris,descriptive_groupby(x= Sepal.Length, group= Species)),
      with(iris,descriptive_groupby(x= Sepal.Width, group= Species)),
      with(iris,descriptive_groupby(x= Petal.Length, group= Species)),
      with(iris,descriptive_groupby(x= Petal.Width, group= Species)))

}
