% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{freq_table}
\alias{freq_table}
\title{Frequency table of a categorical variable.}
\usage{
freq_table(
  x,
  xname = deparse(substitute(x)),
  conf.Int. = FALSE,
  conf.level = 0.95
)
}
\arguments{
\item{x}{must be a factor}

\item{xname}{Name of the variable. Default will be the x name.}

\item{conf.Int.}{Logical argument to calculate the confidence interval of the percentages. Default is \code{FALSE}.}

\item{conf.level}{integer argument, the level of confidence. Default is 0.95,i.e., 95% level of confidence.}
}
\description{
Creates frequency frequencies and percentages of the different levels of a factor variable.
}
\examples{

## iris data:

with(iris,freq_table(x = Species,conf.Int. = TRUE))

}
