% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{indep_two_Populations}
\alias{indep_two_Populations}
\title{Compare a numaric variable between two levels of a factor variable}
\usage{
indep_two_Populations(
  x,
  group,
  basegroup = c(1, 2),
  xname = deparse(substitute(x)),
  plot = FALSE,
  shp = FALSE
)
}
\description{
Compare a numaric variable between two levels of a factor variable
}
