% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{indp_K_Populations}
\alias{indp_K_Populations}
\title{Association of a numeric variable between multiple category (>2) of a factor variable.}
\usage{
indp_K_Populations(x, group, xname = deparse(substitute(x)))
}
\arguments{
\item{x}{must be numeric vector.}

\item{group}{must be factor with greater than 2 levels.}

\item{xname}{Name of numeric variable. Default is the name of x.}
}
\value{
N = Number of observations in \code{x}.

Total= Sum of \code{x}.

MEAN= Mean of \code{x}.

SD= Standard deviation of \code{x}.

MEDIAN= Median of \code{x}.

IQR= Interquartile range (Q1,Q3) of \code{x}.i.e., 1st quantile and the 3rd quantile.

shapiro.test.PValue= P-value of the Shapiro test.

Test.Stat = 1 way ANOVA F statistics.

Par.Pval= Parametric test's P-value,i.e., 1 way ANOVA test's P-value.

NP.Pval= Non parametric test P-value,i.e., Kruskal Wallis test's P-value.

Note: (1) All the summary statistics will be calculated after removing the \code{NA} values.

      (2) If shapiro.test$p-value >0.05 i.e not significant => the data is Normal otherwise,
          <0.05 i.e., significant then it is skewed
}
\description{
It compares the distributions of a numeric variable at different levels of a factor variable and also checks their association
by performing One-way ANOVA if the distribution of the normal or by performing Kruskal Wallis test if it is skewed and summarize
the summary in a data frame format. By default it performs both.
}
\examples{

## IRIS data:

# descriptive summary of  Sepal Length, Sepal Width, Petal Length and, Petal Width

rbind(with(iris,indp_K_Populations(x= Sepal.Length, group= Species)),
      with(iris,indp_K_Populations(x= Sepal.Width, group= Species)),
      with(iris,indp_K_Populations(x= Petal.Length, group= Species)),
      with(iris,indp_K_Populations(x= Petal.Width, group= Species)))

}
