% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{missing.count}
\alias{missing.count}
\title{Column-wise count of missing values in a data frame.}
\usage{
missing.count(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
data.frame
}
\description{
Column-wise count of missing values in a data frame.
}
\examples{

## airquality Data from R:

missing.count(airquality)

}
