% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{mytab_two_two_Contingency}
\alias{mytab_two_two_Contingency}
\title{Association between two categorical variables}
\usage{
mytab_two_two_Contingency(
  x,
  y,
  margin,
  xlev = deparse(substitute(x)),
  Simulate.P.Val = FALSE,
  yets.correction = FALSE,
  plot = FALSE,
  exact.test = FALSE
)
}
\description{
Association between two categorical variables
}
