% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{paired_one_Populations}
\alias{paired_one_Populations}
\title{One sample pare-wise comparison test}
\usage{
paired_one_Populations(x.before, x.after, var)
}
\arguments{
\item{x.before}{measurement of the random variable at the 1st time point or at the 1st condition.}

\item{x.after}{measurement of the random variable at the 2nd time point or at the 2nd condition.}

\item{var}{Name of numeric variable.}
}
\value{
N = Number of observations in \code{x}.

Total= Sum of \code{x}.

MEAN= Mean of \code{x}.

SD= Standard deviation of \code{x}.

MEDIAN= Median of \code{x}.

IQR= Interquartile range (Q1,Q3) of \code{x}.i.e., 1st quantile and the 3rd quantile.

Mean.diff= Mean difference between \code{x.before} and \code{x.after}.

Conf.Int = 95% confidence interval of the mean difference. derived from the paired t-test.

Par.Pval= Parametric test's P-value,i.e., paired t-test's P-value.

NP.Pval= Non parametric test P-value,i.e., Wilcoxon Signed Rank test's P-value.

Note: All the summary statistics will be calculated after removing the \code{NA} values.
}
\description{
This function helps to compare the distribution of two samples when each observation
in one sample can be paired with an observation in the other sample i.e., a measurement
is taken on a subject before and after some treatment or,measurement is taken under two
different conditions – e.g. the response time of a patient is measured on two different drugs.
}
\examples{

## sleep data:

paired_one_Populations(x.before = sleep$extra[sleep$group==1],
                       x.after = sleep$extra[sleep$group==2],
                       var = "extra")

}
