% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{pairwise_Comparison_for_K_Populations}
\alias{pairwise_Comparison_for_K_Populations}
\title{More than two sample pare-wise comparison test}
\usage{
pairwise_Comparison_for_K_Populations(x, group, dunn.method = "bh")
}
\arguments{
\item{x}{must be numeric.}

\item{group}{must be factor with greater than 2 levels.}

\item{dunn.method}{adjusts the p-value for multiple comparisons using the
Bonferroni, Šidák, Holm, Holm-Šidák, Hochberg, Benjamini-Hochberg,
or Benjamini-Yekutieli adjustment (see details: ?FSA::dunnTest). default
option that is used here is "bh"=>Benjamini-Hochberg}
}
\value{
N = Number of observations in \code{x}.

Total= Sum of \code{x}.

MEAN= Mean of \code{x}.

SD= Standard deviation of \code{x}.

MEDIAN= Median of \code{x}.

IQR= Interquartile range (Q1,Q3) of \code{x}.i.e., 1st quantile and the 3rd quantile.

Mean.diff= Mean difference between \code{x.before} and \code{x.after}.

DunnTest_P.adj = Dunn test's adjusted P-value by using bh method.

DunnTest_P.unadj= Dunn test's adjusted P-value to determine statistical significance for each pairwise comparison.
}
\description{
This function helps to perform parametric TukeyHSD post-hoc test and  non-parametric Dunn test
for more than 2 independent groups, to check their association.
}
\note{
Terms like "unadjusted p-value" and "adjusted p-value" are frequently used in statistical hypothesis testing,
especially when performing multiple comparisons.

The p-value derived from a statistical test without multiple comparison adjustment is known as the "unadjusted p-value."
Under the presumption that the null hypothesis is true, it expresses the likelihood of getting a test statistic that is as
extreme as or more extreme than the one that was observed.

For several comparisons done at the same time, the adjusted p-value (also known as the corrected p-value) is utilized.
The probability of incorrectly rejecting the null hypothesis (Type I error) rises with more comparisons. In order to
account for this, adjusted p-values regulate the total Type I error rate (which includes the false discovery rate and
family-wise error rate).

Now, in terms of the Dunn test—a non-parametric substitute for the one-way ANOVA—it is utilized for pairwise comparisons
after the Kruskal-Wallis test. The p-value that is acquired straight out of the Dunn test, without any multiple comparison
correction, is known as the unadjusted p-value.

Here’s what the unadjusted p-value means in the context of the Dunn test:

(1) Each pairwise comparison will produce an uncorrected p-value of its own if you use the Dunn test to perform numerous pairwise
comparisons after discovering a significant result in the Kruskal-Wallis test (which determines whether the distributions of
various groups differ).

(2) A pairwise difference can be deemed statistically significant if its unadjusted p-value is compared to your selected significance
threshold, which is often 0.05. You reject the null hypothesis and determine that there is a significant difference between the
groups under comparison if the unadjusted p-value is smaller than your significance level.

(3) sConducting numerous comparisons raises the likelihood of making a Type I error since each pairwise comparison has its own
unadjusted p-value. Therefore, in order to account for this inflation in the Type I error rate overall, some researchers
decide to modify these p-values.
}
