% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{freq_table_group_by}
\alias{freq_table_group_by}
\title{Group-wise frequency table of a categorical variable}
\usage{
freq_table_group_by(
  Exposure,
  margin = 1,
  Outcome,
  conf.Int. = FALSE,
  conf.level = 0.95
)
}
\description{
Group-wise frequency table of a categorical variable
}
